alteracoes em lab do lab 2

+++ lab-2-keil/src/timer_implementation.h
@@ -9,7 +9,7 @@
 #include "driverlib/sysctl.h"
 #include "driverlib/timer.h"
 
-extern uint8_t freqCarry;
+extern uint32_t freqCarry;
 extern uint32_t freqMeasure;
 extern uint32_t timerCount;
